moved logger to new file, simplified encode queue
