+++ src/index.ts
@@ -1,41 +1,10 @@
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import winston, {createLogger} from 'winston';
-const {combine, timestamp, printf} = winston.format;
+
 import chok from 'chokidar';
 import {tryEncode} from './tryEncode';
 import {Queue} from './Queue';
+import {logger} from './logger';
 
-// ANCHOR: Logging
-
-// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-const textLogFormat = printf(({level, message, timestamp}) => `${timestamp} ${level}: ${message}`);
-
-export const logger = createLogger({
-	level: 'info',
-	format: winston.format.json(),
-	transports: [
-		//
-		// Write all logs with importance level of `error` or less to `error.log`
-		// Write all logs with importance level of `info` or less to `combined.log`
-		//
-		new winston.transports.File({filename: 'logs/error.log', level: 'error'}),
-		new winston.transports.File({filename: `logs/${new Date().getFullYear()}-${new Date().getMonth()}-combined.log`, format: combine(
-			timestamp(),
-			textLogFormat,
-		)}),
-		// Write all logs to console
-		new winston.transports.Console({format: combine(
-			timestamp({format: 'YYYY-MM-DD HH:mm:ss Z'}),
-			winston.format.cli(),
-			textLogFormat,
-		),
-
-		}),
-	],
-});
-
-export const encodeQueue = new Queue({file: './queue.json'});
+const encodeQueue: Queue = new Queue({file: './queue.json'});
 
 // Setup for the watch folder
 // TODO: add an option for the input directory to the config file
@@ -59,7 +28,8 @@ chok.watch('./video-input').on('add', (path: string) => {
 
 // ANCHOR: Encode
 
-tryEncode();
+// Attempts to run encoder function on queue
+tryEncode(encodeQueue);
 
 // ANCHOR: Upload
